% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_darwin_theme.R
\name{gg_darwin_theme}
\alias{gg_darwin_theme}
\title{Title}
\usage{
gg_darwin_theme(plot, text_size = 12, title_size = 16)
}
\arguments{
\item{plot}{placeholder}

\item{text_size}{placeholder}

\item{title_size}{placeholder}
}
\value{
placeholder
}
\description{
Title
}
