% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_darwin_time_series.R
\name{gg_darwin_time_series}
\alias{gg_darwin_time_series}
\title{Title}
\usage{
gg_darwin_time_series(time, start, end, interval)
}
\arguments{
\item{time}{placeholder}

\item{start}{placeholder}

\item{end}{placeholder}

\item{interval}{placeholder}
}
\value{
placeholder
}
\description{
Title
}
