% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_x_date_darwin.R
\name{scale_x_date_darwin}
\alias{scale_x_date_darwin}
\title{Title}
\usage{
scale_x_date_darwin(time, start, end, interval)
}
\arguments{
\item{time}{placeholder}

\item{start}{placeholder}

\item{end}{placeholder}

\item{interval}{placeholder}
}
\value{
placeholder
}
\description{
Title
}
